Add types and functions to process season and streaks
First a function to process the raw game logs into an in-order vector of
games for each team in the season. The other function will take that
list of games and create streaks out of it. Streaks can be of any length
but must contain all of the games of the same type.