Rename RetrosheetGameLog to GameLog
Including the path hierarchy made it retrosheet:games:RetrosheetGameLog.
Obviously it's part of Retrosheet so remove it from the struct name too.

This may get renamed further, but it's already the Chadwick CSV version
of the Retrosheet game log. In the future, something for the Retrosheet
files may be added.