First pass at win/loss palindrome in Rust
The goal of this repository was to have a library and other functions
for processing Retrosheet data in Rust. The win/loss palindrome is
self-contained enough and simple enough that translating the Python
program into Rust was straightforward. Still more to do but the core
works at identifying the games.