Add a "raw" version of GameLog where no Strings are allocated
The GameLog struct has 98 fields with a String type, making for a lot of
allocations if only a couple fields are used. Create a &str version,
GameLogRaw, to allocate none of them and allow the user to decide what
fields need to actually be String. The Option fields may be changed in
the future too because there are 59 of them.